refactor(workflows): neutralize automation brief and volume filenames
