+++ turbo/apps/api/src/signals/routes/zero-workflows.ts
@@ -52,7 +52,7 @@ import {
 import { updateZeroWorkflow$ } from "../services/zero-workflow-update.service";
 import { detectWorkflowConnectorReadiness$ } from "../services/zero-workflow-connector-readiness.service";
 import { createUserMessageDocument } from "../services/zero-chat-user-message.service";
-import { loadWorkflowVolumeFiles } from "../services/zero-workflow-volume.service";
+import { loadWorkflowVolumeFiles } from "../services/workflow-volume.service";
 import {
   encryptWorkflowWebhookSecret,
   encryptWorkflowWebhookToken,
