+++ turbo/apps/api/src/signals/services/workflow-automation-brief.service.ts
